📚 Added Courses Arabic Page
Most of the course names are taken from the official 2018 board which can be found here:
http://gradform.eng.cu.edu.eg/EngDalel/44.htm

N.B: Wasn't sure if it was best to stick to Classical Arabic or Egyptian Arabic or Arabic mixed with English. So, I went with a mix of all 3 while keeping the English to a bare minimum since we already have an English option in the first place.